Fixed compiler warning in crosscorrelator.cpp:1315, the increment of nHits was moved from hitfinder.cpp to worker.cpp so that it displays the correct value, removed the tags in the file names of events for the different hitfinders (they all now save with the standard eventname), changed r0XXX-cleaned.txt to r0XXX-cleanedhits.txt, improved structure of log files so that they (just as the terminal output) show the correct nPeaks for each hitfinder

+++ crosscorrelator.cpp
@@ -1312,7 +1312,7 @@ int CrossCorrelator::calculatePolarCoordinates_FAST( array1D* cartesian1D, array
 			int p_max = maskrow->size();
 			
 			for(int p_ct = 0; p_ct < p_max; p_ct++ ){
-				int maskval = maskrow->get(p_ct);
+				int maskval = (int) maskrow->get(p_ct);
 				if (maskval == 0){								//inside masked out region
 					if ( p_ct == 0 ){							//this row begins with mask --> this is a start
 						start_m.push_back(p_ct);
